+++ src/main/java/com/management/project/controller/ProjectController.java
@@ -43,24 +43,24 @@ public class ProjectController {
 
     @GetMapping("/projects/{id}")
     @Operation(summary = "Finds a project by ID")
-    public ResponseEntity<ProjectResponse> findById(@PathVariable Long id) {
+    public ResponseEntity<ProjectResponse> findById(@PathVariable Integer id) {
         return ResponseEntity.ok(projectService.findById(id));
     }
 
     @GetMapping("/companies/{companyId}/projects")
-    public ResponseEntity<List<ProjectResponse>> findByCompanyId(@PathVariable Long companyId) {
+    public ResponseEntity<List<ProjectResponse>> findByCompanyId(@PathVariable Integer companyId) {
         return ResponseEntity.ok(projectService.findByCompanyId(companyId));
     }
 
     @GetMapping("/persons/{personId}/projects")
-    public ResponseEntity<List<ProjectResponse>> findByPersonId(@PathVariable Long personId) {
+    public ResponseEntity<List<ProjectResponse>> findByPersonId(@PathVariable Integer personId) {
         return ResponseEntity.ok(projectService.findByPersonId(personId));
     }
 
     @GetMapping("/projects/{projectId}/financial-summary")
     @Operation(summary = "Gets the financial summary for a project")
     @PreAuthorize("hasAnyRole('Owner', 'Manager')")
-    public ResponseEntity<ProjectFinancialSummaryDTO> getFinancialSummary(@PathVariable Long projectId) {
+    public ResponseEntity<ProjectFinancialSummaryDTO> getFinancialSummary(@PathVariable Integer projectId) {
         return ResponseEntity.ok(projectFinancialService.getProjectFinancialSummary(projectId));
     }
 }
